code refactoring completed
