+++ app/core.py
@@ -1,25 +1,186 @@
 import os
-import shutil
 import time
 from datetime import datetime
 from threading import Event
 
 import clr
-from PyQt6.QtCore import (Qt, QThreadPool)
+from PyQt6.QtCore import QThreadPool
 from PyQt6.QtGui import QIcon
-from PyQt6.QtWidgets import QFileDialog, QMainWindow
+from PyQt6.QtWidgets import QMessageBox
 
-from downloader import Downloader
-from get_url import url_window
-from misc import DilalogBox, Miscellaneous,Worker
-from url_gen import url_generator
-from utls import open_browser
+from utls import Worker
+from modules.downloader import Downloader
+from modules.gui import Ui_MainProgram
+from modules.url_gen import url_generator
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
 dll_path = os.path.join(script_dir,r"data\System.Management.Automation.dll")
 clr.AddReference(dll_path)
 
-class App(Miscellaneous):
+
+class internal_func(Ui_MainProgram):
+
+    def error_msg(self, text, msg_details, title="Error", critical=False):
+        msg = QMessageBox()
+        msg.setWindowTitle(title)
+        msg.setText(f'{str(text)}     ')
+        if critical:
+            msg.setIcon(QMessageBox.Icon.Critical)
+            msg.setWindowIcon(QIcon('./data/images/error_r.png'))
+        else:
+            msg.setIcon(QMessageBox.Icon.Warning)
+            msg.setWindowIcon(QIcon('./data/images/error_y.png'))
+        msg.setDetailedText(str(msg_details) +
+                            '\n\ncheck Full Logs [Help --> Open Logs]')
+        if text == "Failed To Clear Cache Files!":
+            pass
+        else:
+            self.set_bar_0()
+            self.show_bar(False)
+            self.stop_btn.hide()
+            self.pushButton.setEnabled(True)
+            self.menuDependencies.setEnabled(True)
+            self.actionclear_cache.setEnabled(True)
+            self.actioninstall_From_File.setEnabled(True)
+            self.actionInstall_using_url.setEnabled(True)
+            self.pushButton.show()
+        msg.exec()
+
+    def show_error_popup(self, txt="An Error Has Occured Try Again!"):
+        msg = QMessageBox()
+        msg.setWindowTitle('Error')
+        msg.setWindowIcon(QIcon('./data/images/error_r.png'))
+        msg.setText(f'{txt}     ')
+        msg.setIcon(QMessageBox.Icon.Critical)
+        if txt in ("No Logs Found!", "No Downloads Found!"):
+            pass
+        else:
+            self.set_bar_0()
+            self.show_bar(False)
+            self.stop_btn.hide()
+            self.pushButton.setEnabled(True)
+            self.menuDependencies.setEnabled(True)
+            self.actionclear_cache.setEnabled(True)
+            self.actioninstall_From_File.setEnabled(True)
+            self.actionInstall_using_url.setEnabled(True)
+            self.pushButton.show()
+        msg.exec()
+
+    def show_success_popup(self, text=None):
+        msg = QMessageBox()
+        msg.setWindowTitle('Success')
+        msg.setWindowIcon(QIcon('./data/images/success.png'))
+        if text:
+            msg.setText(f'{text}     ')
+        else:
+            msg.setText('Installation completed!     ')
+        msg.setIcon(QMessageBox.Icon.Information)
+
+        if text == "Cache Files Cleared Successfully!":
+            print("Cache Files Cleared")
+        else:
+            self.set_bar_0()
+            self.show_bar(False)
+            self.stop_btn.hide()
+            self.pushButton.setEnabled(True)
+            self.menuDependencies.setEnabled(True)
+            self.actionclear_cache.setEnabled(True)
+            self.actioninstall_From_File.setEnabled(True)
+            self.actionInstall_using_url.setEnabled(True)
+            self.pushButton.show()
+        msg.exec()
+
+    def error_handler(self, n, normal=True, msg=None, critical=True):
+        def log_error():
+            # if path exits or not
+            if os.path.exists('log.txt'):
+                mode = 'a'
+            else:
+                mode = 'w'
+            # write to the log file
+            with open('log.txt', mode) as f:
+                current_time = datetime.now().strftime("[%d-%m-%Y %H:%M:%S]")
+                f.write(f'[python logs] \n{current_time}\n\n')
+                f.write(n[2])
+                f.write(f'{82*"-"}\n')
+
+        # if normal show a simple popup
+        if normal:
+            log_error()
+            self.show_error_popup()
+        else:
+            msg_details = f'{n[1]}'
+            if msg_details == 'Stoped By User!':
+                self.show_success_popup("Download Stopped!")
+            elif msg_details == 'server returned a empty list':
+                msg = 'Sorry, Application not found!'
+                msg_details = "Application not found in the server, Application is not supported!"
+                self.error_msg(msg, msg_details, "Error", critical)
+            else:
+                log_error()
+                if msg is None:
+                    msg = 'An Error Has Occured Try Again!'
+                self.error_msg(msg, msg_details, "Error", critical)
+
+    def run_success(self, value):
+        if value == 0:
+            self.show_success_popup()
+        else:
+            self.error_msg(*value)
+
+    def main_Progress(self, n):
+        total = self.mainprogressBar.value()
+        if total + n < 100:
+            total += n
+        else:
+            total = 100
+        self.mainprogressBar.setValue(total)
+
+    def cur_Progress(self, n):
+        self.currentprogressBar.setValue(n)
+
+    def set_bar_0(self):
+        self.mainprogressBar.setValue(0)
+        self.currentprogressBar.setValue(0)
+
+    def show_bar(self, val=True):
+        if val is False:
+            self.currentprogressBar.hide()
+            self.mainprogressBar.hide()
+            self.Current_bar.hide()
+            self.Main_bar.hide()
+        elif val is True:
+            self.currentprogressBar.show()
+            self.mainprogressBar.show()
+            self.Current_bar.show()
+            self.Main_bar.show()
+
+    def stop_func(self):
+        self.stop.set()
+        self.stop_btn.hide()
+        self.pushButton.show()
+
+    def closeEvent(self, event):
+        close = QMessageBox()
+        close.setWindowTitle("Confirm")
+        close.setWindowIcon(QIcon('./data/images/error_y.png'))
+        close.setText("Are you sure you want to exit?     ")
+        close.setIcon(QMessageBox.Icon.Warning)
+        close.setStandardButtons(
+            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.Cancel)
+        close = close.exec()
+
+        if close == QMessageBox.StandardButton.Yes:
+            try:
+                self.window.close()
+            except:
+                pass
+            event.accept()
+        else:
+            event.ignore()
+
+
+class core(internal_func):
     def __init__(self):
         super().__init__()
         self.threadpool = QThreadPool()
@@ -190,146 +351,3 @@ class App(Miscellaneous):
         worker.signals.error.connect(
             lambda arg: self.error_handler(arg, normal=False))
         self.threadpool.start(worker)
-
-
-class MainWindowGui(App):
-    def __init__(self):
-        super().__init__()
-
-    def setupUi(self, *args, **kwargs):
-        Miscellaneous.setupUi(self, *args, **kwargs)
-        self.set_bar_0()
-        self.show_bar(False)
-        self.pushButton.clicked.connect(self.openWindow)
-        self.stop_btn.clicked.connect(self.stop_func)
-        self.stop_btn.hide()
-        self.actioninstall_From_File.triggered.connect(
-            self.standalone_installer)
-        self.actionclear_cache.triggered.connect(self.clear_cache)
-        self.actionCheck_For_Updates.triggered.connect(lambda: open_browser(
-            'https://github.com/m-jishnu/alt-app-installer/releases'))
-        self.actionAbout.triggered.connect(lambda: open_browser(
-            'https://github.com/m-jishnu/alt-app-installer'))
-        self.actionHelp.triggered.connect(lambda: open_browser(
-            'https://discord.com/invite/9eeN2Wve4T'))
-        self.actionOpen_Logs.triggered.connect(self.open_Logs)
-        self.actionDownloads.triggered.connect(self.open_downloads)
-        self.actionIgnore_Latest_Version.triggered.connect(self.ignore_version)
-        self.actionIgnore_All_filters.triggered.connect(
-            self.ignore_All_filters)
-        self.actionInstall_using_url.triggered.connect(self.install_url)
-
-    def ignore_version(self):
-        if self.actionIgnore_Latest_Version.isChecked():
-            self.ignore_ver = True
-        else:
-            self.ignore_ver = False
-
-    def ignore_All_filters(self):
-        if self.actionIgnore_All_filters.isChecked():
-            self.all_dependencies = True
-            self.actionIgnore_Latest_Version.setChecked(True)
-            self.actionIgnore_Latest_Version.setEnabled(False)
-        else:
-            self.all_dependencies = False
-            self.actionIgnore_Latest_Version.setChecked(False)
-            self.actionIgnore_Latest_Version.setEnabled(True)
-
-    def open_Logs(self):
-        path = 'log.txt'
-        if os.path.exists(path):
-            os.startfile(path)
-        else:
-            self.show_error_popup(txt="No Logs Found!")
-
-    def clear_cache(self):
-        def remove_file():
-            def remove_(path, mode='file'):
-                if mode == 'file':
-                    if os.path.exists(path):
-                        os.remove(path)
-                    else:
-                        pass
-
-                elif mode == 'dir':
-                    shutil.rmtree(path)
-
-            remove_('log.txt')
-            try:
-                remove_('downloads', 'dir')
-            except FileNotFoundError:
-                print("No Downloads Found!")
-
-        worker = Worker(lambda *ars, **kwargs: remove_file())
-        worker.signals.error.connect(lambda arg: self.error_handler(
-            arg, normal=False, msg="Failed To Clear Cache Files!", critical=False))
-
-        self.threadpool.start(worker)
-        worker.signals.result.connect(lambda: self.show_success_popup(
-            text="Cache Files Cleared Successfully!"))
-
-    def open_downloads(self):
-        path = os.path.realpath("./downloads")
-        if os.path.exists(path):
-            os.startfile(path)
-        else:
-            self.show_error_popup(txt="No Downloads Found!")
-
-    # standalone installer for predownloaded files
-    def standalone_installer(self):
-        def error(arg):
-            self.pushButton.setEnabled(True)
-            self.show_bar(False)
-
-        fname = QFileDialog.getOpenFileNames()[0]
-        if fname:
-            worker = Worker(lambda **kwargs: self.install(fname[0], **kwargs))
-            worker.signals.cur_progress.connect(self.cur_Progress)
-            worker.signals.main_progress.connect(self.main_Progress)
-            worker.signals.result.connect(self.run_success)
-            worker.signals.error.connect(error)
-            self.threadpool.start(worker)
-            self.show_bar(True)
-            self.pushButton.setEnabled(False)
-            self.menuDependencies.setEnabled(False)
-            self.actionclear_cache.setEnabled(False)
-            self.actioninstall_From_File.setEnabled(False)
-            self.actionInstall_using_url.setEnabled(False)
-            # if the app selector window is open closing it
-            try:
-                self.window.close()
-                self.window.deleteLater()
-                del self.window
-            except:
-                pass
-
-    def install_url(self):
-        window = DilalogBox()
-        window.closed.connect(self.parser)
-        window.exec()
-
-    def openWindow(self):
-        # close event for the new window
-        def close(event):
-            self.window.deleteLater()
-            del self.window
-            event.accept()
-
-        try:
-            self.window  # checking if self.window already exist
-        except:
-            self.window = False  # if not set it to false aka the window is not open
-
-        if self.window:  # if it has value then change focus to the already open window
-            self.window.setWindowState(self.window.windowState(
-            ) & ~Qt.WindowState.WindowMinimized | Qt.WindowState.WindowActive)  # if minimized then unminimize
-            self.window.activateWindow()  # set focus to the currently open window
-        else:  # open a new window
-            self.window = QMainWindow()
-            self.window.setWindowIcon(QIcon('./data/images/search.png'))
-            search_app = url_window()
-            search_app.setupUi(self.window)
-            # overding the new window close event for proper cleanup
-            self.window.closeEvent = close
-            self.window.show()
-            search_app.closed.connect(self.parser)
