+++ app/modules/get_url.py
@@ -3,8 +3,8 @@ from PyQt6.QtCore import QObject, QUrl, pyqtSignal
 from PyQt6.QtGui import QAction, QIcon
 from PyQt6.QtWebEngineCore import QWebEnginePage
 from PyQt6.QtWebEngineWidgets import QWebEngineView
-from PyQt6.QtWidgets import (QApplication, QLabel, QLineEdit, QMainWindow,
-                             QMenu, QPushButton, QStatusBar, QToolBar)
+from PyQt6.QtWidgets import (QLabel, QLineEdit, QMenu, QPushButton, QStatusBar,
+                             QToolBar)
 
 
 class CustomWebEngineView(QWebEngineView):
@@ -212,7 +212,7 @@ class url_window(QObject):
         qt_window.urlbar.returnPressed.connect(navigate_to_url)
 
 #test code
-
+# from PyQt6.QtWidgets import (QApplication, QMainWindow)
 # def url_grabber():
 #     import sys
